requirements.txt
torch
torchvision
ultralytics
gradio
pandas
Pillow
